add deck class
