+++ src/main/kotlin/hwr/oop/Game.kt
@@ -1,7 +1,6 @@
 package hwr.oop
 
 class Game {
-    private var deck = ArrayList<Card>()
     private var round: Round = Round()
     private var players = ArrayList<Player>()
     private var dealer: Int = 0
@@ -12,27 +11,6 @@ class Game {
 
     fun players(): List<Player> {
         return players
-    }
-
-    fun resetDeck() {
-        for(suit in CardSuit.entries){
-            for (rank in CardRank.entries){
-                deck += Card(rank, suit)
-            }
-        }
-        shuffleDeck()
-    }
-
-    fun deck(): List<Card> {
-        return deck
-    }
-
-    fun drawCard(): Card {
-        return deck.removeAt(0)
-    }
-
-    fun shuffleDeck() {
-        deck.shuffle()
     }
 
     fun nextDealer() {
