+++ src/main/kotlin/hwr/oop/Round.kt
@@ -1,6 +1,7 @@
 package hwr.oop
 
 class Round {
+    private var deck: Deck = Deck()
     private var cards = ArrayList<Card>()
     private var round: Int = 0
     private var pot: Int = 0
